requirements.txt
-f https://storage.googleapis.com/jax-releases/jax_releases.html
-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html

absl-py==2.2.2
array_record==0.5.1
astunparse==1.6.3
biopython==1.85
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
chex==0.1.6
click==8.1.8
cloudpickle==3.1.1
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
distrax==0.1.2
dm-env==1.6
dm-haiku==0.0.9
dm-tree==0.1.8
enn @ git+https://github.com/deepmind/enn@8be970ae709b27beb4a2b3a4d0a63d4d74d6d9fb
etils==1.5.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.6.0
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.3.25
jaxlib==0.3.25+cuda11.cudnn805
jaxline==0.0.8
jmp==0.0.4
joblib==1.4.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
mizani==0.11.4
ml_collections==0.1.1
ml_dtypes==0.5.1
numpy==1.22.4
oauthlib==3.2.2
opt_einsum==3.4.0
optax==0.1.3
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotnine==0.13.6
promise==2.3
protobuf==3.20.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlax==0.1.6
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.9.3
six==1.17.0
statsmodels==0.14.4
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.1
tensorflow-probability==0.15.0
termcolor==3.0.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
